+++ contracts/Strategies/interfaces/aave/V2/IPriceOracleGetter.sol
@@ -23,5 +23,5 @@ interface IPriceOracleGetter {
    * @param asset the address of the asset
    * @return the ETH price of the asset
    **/
-  function getAssetPrice(address asset) external view returns (uint256);
+  function getAssetPrice(address asset) external view returns (uint);
 }
